requirements.txt.txt
fastapi
uvicorn[standard]
tortoise-orm
jinja2
htmx
aiofiles
pyyaml
